requirements.txt
beautifulsoup4>=4.12.3
requests>=2.32.3
lxml>=5.2.2
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1
rank-bm25>=0.2.2
numpy>=1.26.4
tqdm>=4.66.4
playwright>=1.46.0
